Allows recalculation of patient caches
In order to recalculate patient caches we want to use the recalculate
option in the calculation job. Using the recalculate flag we can force
an expiration of results.

We also added an expired_at value to the patient caches in order for
them not to be included in future calculations and results. These can be
purged at any time in the future.